added shopped functionality

+++ app/forms/pantry_ingredients.py
@@ -1,4 +1,4 @@
-from wtforms import IntegerField, Form
+from wtforms import IntegerField, Form, FloatField
 from wtforms.validators import DataRequired
 
 
@@ -6,4 +6,4 @@ class PantryIngredientForm(Form):
     user_id = IntegerField('userId', validators=[DataRequired()])
     ingredient_id = IntegerField('ingredientId', validators=[DataRequired()])
     measurement_id = IntegerField('measurementId', validators=[DataRequired()])
-    quantity = IntegerField('quantity', validators=[DataRequired()])
+    quantity = FloatField('quantity', validators=[DataRequired()])
